+++ oneline/oneline/Models/Score.cs
@@ -12,7 +12,7 @@ namespace oneline.Models
         public int ScoreIdx { get; set; }
         public string UserId { get; set; }
         public int WorldIdx { get; set; }
-        public int MyScore { get; set; }
+        public float MyScore { get; set; }
         public User User { get; set; }
         public World World { get; set; }
